Treat lineItems as a nested list field
Line items are returned as part of the transaction detail object, and there can be 0 or more line item objects. Current functionality works properly for 0 or 1 line item, but if there are >1 line items the other line items are ignored silently. Simply adding lineItems to this list fixes it for me, and I believe this is the appropriate way to do this (although obviously I defer to the maintainers).